// gatilho/web/src/lib/types.ts
export interface TagItem {
  id: string;
  name: string;
  color: string;
  contactsCount?: number;
}

export interface FieldItem {
  id: string;
  key: string;
  label: string;
  type: string;
  isSystem: boolean;
}